Extract template finder and ensure tests keep working when the templates are not present
